+++ R/global_imports.R
@@ -1,3 +1,0 @@
-#' @importFrom rlang .data .env
-#' @importFrom magrittr %>%
-NULL