+++ NEWS.md
@@ -1,7 +1,12 @@
-# connecttest 0.0.2 *2022-16*
+# connecttest 0.0.3 _2022-09-20_
 
+-   Removes unused imports
+
+# connecttest 0.0.2 _2022-09-16_
+
+-   Added test which deploys an RMarkdown file to Connect
 -   Updated DESCRIPTION
 
-# connecttest 0.0.1 *2021-01-01*
+# connecttest 0.0.1 _2021-01-01_
 
 -   Initial commit